Split up more volumes, move FluidType enum
