+++ NewHorizons/External/Modules/Volumes/VanishVolumeInfo.cs
@@ -1,0 +1,20 @@
+using Newtonsoft.Json;
+using System.ComponentModel;
+
+namespace NewHorizons.External.Modules.Volumes
+{
+    [JsonObject]
+    public class VanishVolumeInfo : VolumeInfo
+    {
+        /// <summary>
+        /// Whether the bodies will shrink when they enter this volume or just disappear instantly.
+        /// </summary>
+        [DefaultValue(true)] public bool shrinkBodies = true;
+
+        /// <summary>
+        /// Whether this volume only affects the player and ship.
+        /// </summary>
+        public bool onlyAffectsPlayerAndShip;
+    }
+
+}
